% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utilities.R
\name{alg_as_logical_char_logical}
\alias{alg_as_logical_char_logical}
\title{Coerce to logical a character vector which is "char_logical"}
\usage{
alg_as_logical_char_logical(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical vector
}
\description{
Coerce to logical a character vector which is "char_logical"
}
