% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utilities.R
\name{alg_is_char_logical}
\alias{alg_is_char_logical}
\title{Check if a vector can be coerced to logical without information loss}
\usage{
alg_is_char_logical(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE if x can be losslessly coerced to logical
}
\description{
Check if a vector can be coerced to logical without information loss
}
