% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utilities.R
\name{alg_make_clean_names}
\alias{alg_make_clean_names}
\title{Transform a character string into a clean name}
\usage{
alg_make_clean_names(x)
}
\arguments{
\item{x}{the original character vector}
}
\value{
a vector of clean names
}
\description{
A clean name is defined here as a valid R name and:
* only ASCII characters
* underscore to separate words instead of dots
* only lowercase
}
