% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utilities.R
\name{alg_map_col_values}
\alias{alg_map_col_values}
\title{Clean all columns with a value map}
\usage{
alg_map_col_values(df, value_maps)
}
\arguments{
\item{df}{the data frame to modify}

\item{value_maps}{the list of mapped values}
}
\value{
the df were columns were modified
}
\description{
Clean all columns with a value map
}
