% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates.R
\name{alg_parse_excel_date}
\alias{alg_parse_excel_date}
\title{Parse an excel date (text or integer) into date}
\usage{
alg_parse_excel_date(e_date)
}
\arguments{
\item{e_date}{the text or integer vector}
}
\value{
a correct date
}
\description{
Excel stores date as days since 1900-01-01
}
