% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expl.R
\name{cc_test}
\alias{cc_test}
\title{categorical - categorical statistical testing}
\usage{
cc_test(df, exed, exry, na.rm = FALSE)
}
\arguments{
\item{exed}{variable to explain}

\item{exry}{explainatory variable}
}
\value{
a data frame containing descriptions and p-values
}
\description{
categorical - categorical statistical testing
}
