% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sens_spe.R
\name{ci_prop}
\alias{ci_prop}
\title{Calculate a proportion with CI95}
\usage{
ci_prop(ts, n, prec)
}
\arguments{
\item{ts}{number of truth}

\item{n}{number of observation}

\item{prec}{the default precision for `label_percent`}
}
\value{
a formatted string of p% [CI95]
}
\description{
Calculate a proportion with CI95
}
