% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expl.R
\name{cq_test}
\alias{cq_test}
\title{categorical - quanti statistical testing}
\usage{
cq_test(df, exed, exry, na.rm = FALSE)
}
\arguments{
\item{exed}{variable to explain}

\item{exry}{explainatory variable}
}
\value{
a data frame containing descriptions and p-values
}
\description{
Reorder the parameters and send to qc_test
}
