% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-roc.R
\name{ggroc_youden}
\alias{ggroc_youden}
\title{Extend the ggroc function to add the Youden best cutoff}
\usage{
ggroc_youden(roc_obj)
}
\arguments{
\item{roc_obj}{}
}
\value{
a ggplot graph
}
\description{
Extend the ggroc function to add the Youden best cutoff
}
