% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-merge_sequences.R
\name{merge_sequences}
\alias{merge_sequences}
\title{Return merged sequences of begining and ends}
\usage{
merge_sequences(beg, end, na.rm = FALSE)
}
\arguments{
\item{beg}{start of each sequence}

\item{end}{end of each sequence}

\item{na.rm}{boolean, do we remove sequences containing NAs?}
}
\value{
a list of 2, beg and end, the merged sequences
}
\description{
Return merged sequences of begining and ends
}
