% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expl.R
\name{qc_test}
\alias{qc_test}
\title{quanti - categorical statistical testing}
\usage{
qc_test(df, exed, exry, na.rm = FALSE)
}
\arguments{
\item{exed}{variable to explain}

\item{exry}{explainatory variable}
}
\value{
a data frame containing descriptions and p-values
}
\description{
quanti - categorical statistical testing
}
