% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expl.R
\name{qq_test}
\alias{qq_test}
\title{quanti - quanti statistical testing}
\usage{
qq_test(df, exed, exry)
}
\arguments{
\item{exed}{variable to explain}

\item{exry}{explainatory variable}
}
\value{
a data frame containing descriptions and p-values
}
\description{
quanti - quanti statistical testing
}
