% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sens_spe.R
\name{senspe_fn}
\alias{senspe_fn}
\title{Make of formatted table of sens, spe, vpp, vpn}
\usage{
senspe_fn(e, m, prec = 0.1)
}
\arguments{
\item{e}{logical vector of the test result (test to eval)}

\item{m}{logical vector of the disease (ground truth)}

\item{prec}{the default precision for `label_percent`}
}
\value{
a formatted tibble
}
\description{
Make of formatted table of sens, spe, vpp, vpn
}
