% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-roc.R
\name{youden_j_max}
\alias{youden_j_max}
\title{Calculate the maximum Youden J index}
\usage{
youden_j_max(roc_obj)
}
\arguments{
\item{roc_obj}{A roc object from pROC::roc}
}
\value{
a vector containing the max Youden index and the specificity and sensitivity corresponding
}
\description{
Calculate the Youden J index, outputs its value and the specificity and sensitivity corresponding
}
